ARM: constify machine_desc structure uses
struct machine_desc records are defined everywhere as a 'const'
structure, but unfortuantely it loses its const-ness through the use of
linker magic - the symbols which surround the section are not declared
const so it becomes possible not to use 'const' for pointers to these
const structures.

Let's fix this oversight - all pointers to these structures should be
marked const too.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
[joonwoop@codeaurora.org: fixed trivial merge conflict. constify some
 more code to avoid warning.]
Signed-off-by: Joonwoo Park <joonwoop@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Dorimanx Note:
also merged this commit here

https://github.com/neobuddy89/shamu/commit/bd51e2f595580fb64bd8494badef746c4d626b25

Conflicts:
	arch/arm/Kconfig
	arch/arm/include/asm/prom.h