proc: uid: Adds accounting for the cputimes per uid.
Adds proc files /proc/uid_cputime/show_uid_stat and
/proc/uid_cputime/remove_uid_range.

show_uid_stat lists the total utime and stime for the active as well as
terminated processes for each of the uids.

Writing a range of uids to remove_uid_range will delete the accounting
for all the uids within that range.

Conflicts:
	drivers/misc/Kconfig
	drivers/misc/Makefile

Change-Id: Ie1ee76caea7ef4f343fd37e0ffa00556ec85eb3a

Conflicts:
	drivers/misc/Kconfig
	drivers/misc/Makefile