Revert " cpufreq: stats: do cpufreq_cpu_put() corresponding to cpufreq_cpu_get()"
This reverts commit 8d245854d058f646db95aa0ce6cd5cea66395ba5.

Dorimanx note:
I have found bug trace to the sysfs free function that managed to work
when there was nothing to free sysfs knob was not created.
so it's was NULL. and fs/sysfs/inode.c crashed with null pointer.

seen its few times. revert to 3.4.y function, as this driver is close to stock.