RESTORE "mm: process_reclaim: fix reclaim skip on low efficiency""
This reverts commit 3542b8d57393ddbf9d820cc3b62b0a53265688e6.

I have made debug, and i have found that driver is not working at all!

the WORK function never start, as it's fail on if (!atomic_dec_and_test(&skip_reclaim))

The skip_reclaim is always negative in range of -3300~-3500
as no one exist to update it, driver not working at all.

reverting this, fix it. and it's reclaiming again as should.

CM13 you are wrong. sorry.